Fixed expanding/collapsing logic & cleanup
